libc: arm64: Use SIMD-accelerated memcpy
It turns out bionic is not using SIMD variant of memcpy although NEON can
be used in userspace. Import memcpy-advsimd from optimized-routines and
use it. Guard was added for extra safety but it is probably superfluous
on every modern Android device.

Initial commit in arm repo stated performance increase of 11%.

Signed-off-by: Andrzej Perczak <linux@andrzejperczak.com>
Signed-off-by: Aaradhay Vijay <aaradhayvijay964@gmail.com>
Change-Id: I1103ce688ecc24bc8a578eed4babd70ae3e95a42
Signed-off-by: Omar Mohamed <omarcoptan9@gmail.com>